+++ lib/size_screen.dart
@@ -11,22 +11,26 @@ class SizeScreen {
     return MediaQuery.of(context).size.height;
   }
 
-  /// Retorna percentual da largura total
+  /// Retorna percentual da largura total.
+  /// percentage vai de 0 à 100
   static double perWidth(BuildContext context, int percentage) {
     return (MediaQuery.of(context).size.width * percentage) / 100;
   }
 
-  /// Retorna percentual da altura total
+  /// Retorna percentual da altura total.
+  /// percentage vai de 0 à 100
   static double perHeight(BuildContext context, int percentage) {
     return (MediaQuery.of(context).size.height * percentage) / 100;
   }
 
-  /// Retorna fração da largura total
+  /// Retorna fração da largura total.
+  /// divisor vai de 0 à 1
   static double fracWidth(BuildContext context, double divisor) {
     return MediaQuery.of(context).size.width / divisor;
   }
 
   /// Retorna fração da altura total
+  /// divisor vai de 0 à 1
   static double fracHeight(BuildContext context, double divisor) {
     return MediaQuery.of(context).size.height / divisor;
   }