+++ test/date_utils_test.dart
@@ -1,7 +1,6 @@
 // ignore_for_file: deprecated_member_use_from_same_package
 
 import 'package:esig_utils/date_utils.dart';
-import 'package:esig_utils/models/dia.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:intl/date_symbol_data_local.dart';
 
@@ -24,25 +23,6 @@ void main() {
   ];
 
   group('EsigDateUtils: ', () {
-    Dia dia;
-    List<Dia> dias;
-
-    test('Extensions', () {
-      dia = Dia(2020, 1, 1);
-      expect(dia.menosUmDia, Dia(2019, 12, 31));
-
-      dia = dia.menosUmDia;
-      expect(dia.maisUmDia, Dia(2020, 1, 1));
-
-      dia = Dia(2020, 10, 25); // domingo
-      var diaUtil = [false, true, true, true, true, true, false];
-
-      for (var du in diaUtil) {
-        expect(dia.util, du);
-        dia = dia.maisUmDia;
-      }
-    });
-
     test('datetimeFromString', () {
       var d = EsigDateUtils.datetimeFromString('01/02/2022');
       expect(d, DateTime(2022, 2, 1));
@@ -53,84 +33,6 @@ void main() {
       d = EsigDateUtils.datetimeFromString('01/02/2022 03:59',
           format: 'dd/MM/yyyy HH:mm');
       expect(d, DateTime(2022, 2, 1, 3, 59));
-    });
-    test('strDMAToDia', () {
-      expect(EsigDateUtils.strDiaMesAnoToDia('02-03-2020'), Dia(2020, 3, 2));
-      expect(EsigDateUtils.strDiaMesAnoToDia('01-02-2021'), Dia(2021, 2, 1));
-    });
-
-    test('getNomeDiaSemana', () {
-      var diaNomes = [
-        'DOMINGO',
-        'SEGUNDA-FEIRA',
-        'TERÇA-FEIRA',
-        'QUARTA-FEIRA',
-        'QUINTA-FEIRA',
-        'SEXTA-FEIRA',
-        'SÁBADO'
-      ];
-      dia = Dia(2020, 10, 25);
-
-      for (var diaNome in diaNomes) {
-        expect(EsigDateUtils.getNomeDiaSemana(dia), diaNome);
-        dia = dia.maisUmDia;
-      }
-    });
-
-    test('getPrimeiroDomingo', () {
-      // Passando dia perto do final do mês
-      dia = Dia(2020, 10, 30);
-      expect(EsigDateUtils.getPrimeiroDomingo(dia), Dia(2020, 9, 27));
-
-      // Passando primeiro dia do mês
-      dia = Dia(2020, 9, 1);
-      expect(EsigDateUtils.getPrimeiroDomingo(dia), Dia(2020, 8, 30));
-
-      // Passando primeiro dia do mês
-      dia = Dia(2020, 10, 1);
-      expect(EsigDateUtils.getPrimeiroDomingo(dia), Dia(2020, 9, 27));
-    });
-
-    test('getUltimoSabado', () {
-      // Passando o próprio resultado
-      dia = Dia(2020, 10, 31);
-      expect(EsigDateUtils.getUltimoSabado(dia), Dia(2020, 10, 31));
-
-      // Passando primeiro dia do mês
-      dia = Dia(2020, 9, 1);
-      expect(EsigDateUtils.getUltimoSabado(dia), Dia(2020, 10, 3));
-
-      // Passando dia perto do final do mês
-      dia = Dia(2020, 10, 30);
-      expect(EsigDateUtils.getUltimoSabado(dia), Dia(2020, 10, 31));
-    });
-
-    test('último dia do mês', () {
-      expect(Dia(2020, 10, 0), Dia(2020, 9, 30));
-      expect(Dia(2020, 11, 0), Dia(2020, 10, 31));
-      expect(Dia(2021, 1, 0), Dia(2020, 12, 31));
-    });
-
-    test('ultimoSabado', () {
-      // Passando dia perto do final do mês
-      dia = Dia(2020, 10, 30);
-      expect(EsigDateUtils.getUltimoSabado(dia), Dia(2020, 10, 31));
-
-      // Passando primeiro dia do mês
-      dia = Dia(2020, 9, 1);
-      expect(EsigDateUtils.getUltimoSabado(dia), Dia(2020, 10, 3));
-
-      // Passando primeiro dia do mês
-      dia = Dia(2020, 10, 1);
-      expect(EsigDateUtils.getUltimoSabado(dia), Dia(2020, 10, 31));
-
-      // Passando último dia do ano 2020
-      dia = Dia(2020, 12, 31);
-      expect(EsigDateUtils.getUltimoSabado(dia), Dia(2021, 1, 2));
-
-      // Passando primeiro dia do ano 2020
-      dia = Dia(2020, 1, 1);
-      expect(EsigDateUtils.getUltimoSabado(dia), Dia(2020, 2, 1));
     });
 
     test('getNomeMeses', () {
@@ -146,32 +48,6 @@ void main() {
               '${nomesMeses[i - 1]} / ${nomesMeses[j - 1]}');
         }
       }
-    });
-
-    test('getListaDias', () {
-      // Passando primeiro e último dia de outubro
-      dias = EsigDateUtils.getListaDias(Dia(2020, 10, 1), Dia(2020, 10, 31));
-      expect(dias.length, 31);
-      expect(dias.first, Dia(2020, 10, 1));
-      expect(dias.last, Dia(2020, 10, 31));
-
-      /// Passando ano bissexto
-      dias = EsigDateUtils.getListaDias(Dia(2020, 1, 1), Dia(2020, 12, 31));
-      expect(dias.length, 366);
-
-      /// Passando ano não bissexto
-      dias = EsigDateUtils.getListaDias(Dia(2019, 1, 1), Dia(2019, 12, 31));
-      expect(dias.length, 365);
-
-      /// Passando um [ultimo] antes de [inicio]
-      dias = EsigDateUtils.getListaDias(Dia(2020, 2, 1), Dia(2020, 1, 1));
-      expect(dias.length, 0);
-
-      /// Função não deve alterar valores de [inicio] e [fim]
-      Dia inicio = Dia(2020, 1, 1), fim = Dia(2020, 2, 31);
-      dias = EsigDateUtils.getListaDias(inicio, fim);
-      expect(inicio, Dia(2020, 1, 1));
-      expect(fim, Dia(2020, 2, 31));
     });
 
     test('Bimestre', () async {